@@ -11,43 +11,43 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.2.9",
-    "@angular/cdk": "^7.3.5",
-    "@angular/common": "^7.2.9",
-    "@angular/compiler": "^7.2.9",
-    "@angular/core": "^7.2.9",
-    "@angular/fire": "^5.1.2",
-    "@angular/forms": "^7.2.9",
-    "@angular/material": "^7.3.5",
-    "@angular/platform-browser": "^7.2.9",
-    "@angular/platform-browser-dynamic": "^7.2.9",
-    "@angular/router": "^7.2.9",
-    "@fortawesome/angular-fontawesome": "^0.3.0",
-    "@fortawesome/fontawesome-svg-core": "^1.2.15",
-    "@fortawesome/free-solid-svg-icons": "^5.7.2",
-    "add": "^2.0.6",
-    "angular-font-awesome": "^3.1.2",
-    "bootstrap": "^4.3.1",
-    "core-js": "^2.6.5",
-    "firebase": "^5.9.0",
-    "font-awesome": "^4.7.0",
+    "@angular/animations": "^8.0.0",
+    "@angular/cdk": "^8.0.0",
+    "@angular/common": "^8.0.0",
+    "@angular/compiler": "^8.0.0",
+    "@angular/core": "^8.0.0",
+    "@angular/fire": "^5.2.1",
+    "@angular/forms": "^8.0.0",
+    "@angular/material": "^8.0.0",
+    "@angular/platform-browser": "^8.0.0",
+    "@angular/platform-browser-dynamic": "^8.0.0",
+    "@angular/router": "^8.0.0",
+    "@fortawesome/angular-fontawesome": "^0.4.0",
+    "@fortawesome/fontawesome-svg-core": "^1.2.19",
+    "@fortawesome/free-brands-svg-icons": "^5.9.0",
+    "@fortawesome/free-regular-svg-icons": "^5.9.0",
+    "@fortawesome/free-solid-svg-icons": "^5.9.0",
+    "@ng-bootstrap/ng-bootstrap": "^4.2.1",
+    "core-js": "^2.6.9",
+    "firebase": "^5.11.1",
+    "hammerjs": "^2.0.8",
     "ngx-ckeditor": "^0.4.0",
-    "ngx-image-cropper": "^1.3.8",
-    "pdfjs-dist": "^2.0.943",
-    "rxjs": "~6.3.3",
+    "ngx-file-drop": "^8.0.2",
+    "ngx-image-cropper": "^1.3.10",
+    "ngx-spinner": "^7.1.4",
+    "rxjs": "~6.5.2",
     "tslib": "^1.9.0",
-    "yarn": "^1.13.0",
-    "zone.js": "^0.8.29"
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.12.4",
-    "@angular/cli": "^7.3.6",
-    "@angular/compiler-cli": "^7.2.9",
-    "@angular/language-service": "^7.2.9",
+    "@angular-devkit/build-angular": "^0.800.2",
+    "@angular/cli": "^8.0.2",
+    "@angular/compiler-cli": "^8.0.0",
+    "@angular/language-service": "^8.0.0",
     "@types/jasmine": "^2.8.16",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "~8.9.4",
-    "codelyzer": "~4.5.0",
+    "codelyzer": "^5.0.1",
     "jasmine-core": "~2.99.1",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "^3.1.4",
@@ -58,6 +58,6 @@
     "protractor": "^5.4.2",
     "ts-node": "~7.0.0",
     "tslint": "~5.11.0",
-    "typescript": "~3.1.6"
+    "typescript": "~3.4.5"
   }
 }
