@@ -1,4 +1,6 @@
-import { Component } from '@angular/core';
+import {Component, Renderer2} from '@angular/core';
+import {NavigationStart, Router} from '@angular/router';
+import {environment} from '../environments/environment';
 
 @Component({
   selector: 'app-root',
@@ -7,4 +9,20 @@ import { Component } from '@angular/core';
 })
 export class AppComponent {
   title = 'Testock';
+  previousUrl: string;
+  production = environment.production;
+  constructor(private renderer: Renderer2, private router: Router) {
+    this.router.events.subscribe((event) => {
+      if (event instanceof NavigationStart) {
+        if (this.previousUrl) {
+          this.renderer.removeClass(document.body, this.previousUrl);
+        }
+        const currentUrlSlug = event.url.slice(1);
+        if (currentUrlSlug) {
+          this.renderer.addClass(document.body, currentUrlSlug);
+        }
+        this.previousUrl = currentUrlSlug;
+      }
+    });
+  }
 }
