@@ -1,4 +1,4 @@
-import { SafeUrlPipe } from './safe-url.pipe';
+import {SafeUrlPipe} from './safe-url.pipe';
 
 describe('SafeUrlPipe', () => {
   it('create an instance', () => {
