@@ -1,6 +1,6 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import {async, ComponentFixture, TestBed} from '@angular/core/testing';
 
-import { FacultiesComponent } from './faculties.component';
+import {FacultiesComponent} from './faculties.component';
 
 describe('FacultiesComponent', () => {
   let component: FacultiesComponent;
