@@ -1,6 +1,6 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import {async, ComponentFixture, TestBed} from '@angular/core/testing';
 
-import { FacultyComponent } from './faculty.component';
+import {FacultyComponent} from './faculty.component';
 
 describe('FacultyComponent', () => {
   let component: FacultyComponent;
