@@ -1,6 +1,6 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import {async, ComponentFixture, TestBed} from '@angular/core/testing';
 
-import { SolutionComponent } from './solution.component';
+import {SolutionComponent} from './solution.component';
 
 describe('SolutionComponent', () => {
   let component: SolutionComponent;
